vvvvv

+++ src/main/java/com/FrameWork/MedLite/Authentification/repository/UserRepository.java
@@ -16,7 +16,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Repository
-public interface UserRepository extends ListCrudRepository<User, Integer> {
+public interface UserRepository extends ListCrudRepository<User, Long> {
 
     Optional<User> findByUserName(String userName);
